feat: redirect policy-repository/ to subjects/
