+++ solution/backend/regulations/templates/regulations/subjects.html
@@ -2,7 +2,7 @@
 {% load static %}
 {% load string_formatters %}
 
-{% block title %}Policy Repository{% endblock %}
+{% block title %}Browse by Subject{% endblock %}
 
 {% block spa_styles %}
     <link href="https://cdn.jsdelivr.net/npm/@mdi/font@4.x/css/materialdesignicons.min.css" rel="stylesheet">
