+++ solution/backend/regulations/views/policy_repository.py
@@ -1,21 +1,8 @@
-from django.views.generic.base import TemplateView
+from django.http import HttpResponsePermanentRedirect
+from django.urls import reverse
 
 
-class PolicyRepositoryView(TemplateView):
-
-    template_name = 'regulations/policy_repository.html'
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        host = self.request.get_host()
-        is_authenticated = self.request.user.is_authenticated
-        user_groups = [group.name for group in self.request.user.groups.all()]
-        has_editable_job_code = any(group in ['EREGS_ADMIN', 'EREGS_MANAGER', 'EREGS_EDITOR'] for group in user_groups)
-
-        c = {
-            'host': host,
-            'is_authenticated': is_authenticated,
-            'has_editable_job_code': has_editable_job_code,
-        }
-
-        return {**context, **c}
+def PolicyRepositoryView(request):
+    redirect_string = '?' + request.GET.urlencode() if request.GET else ''
+    response = HttpResponsePermanentRedirect(reverse('subjects') + redirect_string)
+    return response
